reactor(storage/sql): embed migrations into flipt binary
