+++ config/production.yml
@@ -13,5 +13,3 @@ server:
 
 db:
   url: postgres://postgres@localhost:5432/flipt?sslmode=disable
-  migrations:
-    path: ./config/migrations
