+++ internal/config/testdata/advanced.yml
@@ -32,8 +32,6 @@ tracing:
 
 db:
   url: postgres://postgres@localhost:5432/flipt?sslmode=disable
-  migrations:
-    path: ./config/migrations
   max_idle_conn: 10
   max_open_conn: 50
   conn_max_lifetime: 30m
