+++ internal/config/testdata/default.yml
@@ -25,5 +25,3 @@
 
 # db:
 #   url: file:/var/opt/flipt/flipt.db
-#   migrations:
-#     path: /etc/flipt/config/migrations
