+++ internal/config/testdata/deprecated/database_migrations_path.yml
@@ -1,0 +1,2 @@
+db:
+  migrations_path: "../config/migrations"
