+++ internal/config/testdata/deprecated/database_migrations_path_legacy.yml
@@ -1,0 +1,3 @@
+db:
+  migrations:
+    path: "../config/migrations"
