+++ internal/storage/sql/flag_test.go
@@ -1,4 +1,4 @@
-package sql
+package sql_test
 
 import (
 	"context"
@@ -7,6 +7,7 @@ import (
 	"time"
 
 	"go.flipt.io/flipt/internal/storage"
+	fliptsql "go.flipt.io/flipt/internal/storage/sql"
 	"go.flipt.io/flipt/internal/storage/sql/common"
 	flipt "go.flipt.io/flipt/rpc/flipt"
 
@@ -100,7 +101,7 @@ func (s *DBTestSuite) TestListFlagsPagination_LimitOffset() {
 	}
 
 	for _, req := range reqs {
-		if s.driver == MySQL {
+		if s.db.Driver == fliptsql.MySQL {
 			// required for MySQL since it only s.stores timestamps to the second and not millisecond granularity
 			time.Sleep(time.Second)
 		}
@@ -173,7 +174,7 @@ func (s *DBTestSuite) TestListFlagsPagination_LimitWithNextPage() {
 	}
 
 	for _, req := range reqs {
-		if s.driver == MySQL {
+		if s.db.Driver == fliptsql.MySQL {
 			// required for MySQL since it only s.stores timestamps to the second and not millisecond granularity
 			time.Sleep(time.Second)
 		}
