+++ internal/storage/sql/metrics.go
@@ -2,6 +2,7 @@ package sql
 
 import (
 	"database/sql"
+	"sync"
 
 	"github.com/prometheus/client_golang/prometheus"
 )
@@ -16,63 +17,67 @@ type statsGetter interface {
 	Stats() sql.DBStats
 }
 
+var once sync.Once
+
 // nolint
 func registerMetrics(d Driver, s statsGetter) {
-	labels := prometheus.Labels{"driver": d.String()}
+	once.Do(func() {
+		labels := prometheus.Labels{"driver": d.String()}
 
-	collector := &metricsCollector{
-		sg: s,
-		maxOpenDesc: prometheus.NewDesc(
-			prometheus.BuildFQName(namespace, subsystem, "max_open_conn"),
-			"Maximum number of open connections to the database.",
-			nil,
-			labels,
-		),
-		openDesc: prometheus.NewDesc(
-			prometheus.BuildFQName(namespace, subsystem, "open_conn"),
-			"The number of established connections both in use and idle.",
-			nil,
-			labels,
-		),
-		inUseDesc: prometheus.NewDesc(
-			prometheus.BuildFQName(namespace, subsystem, "in_use_conn"),
-			"The number of connections currently in use.",
-			nil,
-			labels,
-		),
-		idleDesc: prometheus.NewDesc(
-			prometheus.BuildFQName(namespace, subsystem, "idle_conn"),
-			"The number of idle connections.",
-			nil,
-			labels,
-		),
-		waitedForDesc: prometheus.NewDesc(
-			prometheus.BuildFQName(namespace, subsystem, "waited_for_conn"),
-			"The total number of connections waited for.",
-			nil,
-			labels,
-		),
-		blockedSecondsDesc: prometheus.NewDesc(
-			prometheus.BuildFQName(namespace, subsystem, "blocked_seconds_conn"),
-			"The total time blocked waiting for a new connection.",
-			nil,
-			labels,
-		),
-		closedMaxIdleDesc: prometheus.NewDesc(
-			prometheus.BuildFQName(namespace, subsystem, "closed_max_idle_conn"),
-			"The total number of connections closed due to SetMaxIdleConns.",
-			nil,
-			labels,
-		),
-		closedMaxLifetimeDesc: prometheus.NewDesc(
-			prometheus.BuildFQName(namespace, subsystem, "closed_max_lifetime_conn"),
-			"The total number of connections closed due to SetConnMaxLifetime.",
-			nil,
-			labels,
-		),
-	}
+		collector := &metricsCollector{
+			sg: s,
+			maxOpenDesc: prometheus.NewDesc(
+				prometheus.BuildFQName(namespace, subsystem, "max_open_conn"),
+				"Maximum number of open connections to the database.",
+				nil,
+				labels,
+			),
+			openDesc: prometheus.NewDesc(
+				prometheus.BuildFQName(namespace, subsystem, "open_conn"),
+				"The number of established connections both in use and idle.",
+				nil,
+				labels,
+			),
+			inUseDesc: prometheus.NewDesc(
+				prometheus.BuildFQName(namespace, subsystem, "in_use_conn"),
+				"The number of connections currently in use.",
+				nil,
+				labels,
+			),
+			idleDesc: prometheus.NewDesc(
+				prometheus.BuildFQName(namespace, subsystem, "idle_conn"),
+				"The number of idle connections.",
+				nil,
+				labels,
+			),
+			waitedForDesc: prometheus.NewDesc(
+				prometheus.BuildFQName(namespace, subsystem, "waited_for_conn"),
+				"The total number of connections waited for.",
+				nil,
+				labels,
+			),
+			blockedSecondsDesc: prometheus.NewDesc(
+				prometheus.BuildFQName(namespace, subsystem, "blocked_seconds_conn"),
+				"The total time blocked waiting for a new connection.",
+				nil,
+				labels,
+			),
+			closedMaxIdleDesc: prometheus.NewDesc(
+				prometheus.BuildFQName(namespace, subsystem, "closed_max_idle_conn"),
+				"The total number of connections closed due to SetMaxIdleConns.",
+				nil,
+				labels,
+			),
+			closedMaxLifetimeDesc: prometheus.NewDesc(
+				prometheus.BuildFQName(namespace, subsystem, "closed_max_lifetime_conn"),
+				"The total number of connections closed due to SetConnMaxLifetime.",
+				nil,
+				labels,
+			),
+		}
 
-	prometheus.MustRegister(collector)
+		prometheus.MustRegister(collector)
+	})
 }
 
 type metricsCollector struct {
